Refactor createRoundsStatusFilter into utils
